Fixes tracing for the indexing part of quickwit. (#2056)
Before this PR actors would be executed on a root span that was
associated to their execution loop.

That span would be gigantic and would never be closed.

In addition, the message spans did not have any associated name by
default. This solution made logs that looked good in stdout but were
terribly broken when consumed as traces in a tool like jaeger.

The following PR fixes the issue by removing the crazy root span.
It also does something a tiny bit more ambitious by creating a root span
per indexing batch. The batch follows the split, so that it is easier
to inspect how a trip travels through the indexing pipeline.

It is also possible to see the associated context, timings etc.
for each split creation or merge operation.

Closes #2055