Fix broken doc link in quickwit-storage (#3735)
In #3709 there was a refactor that removed `StorageConfig` from the
scope of the changed file in this commit. This was fine as the function
`register` in StorageResolveBuilder did not need it anymore. However,
the docs alluded to registering both a factory and a config. When
running `make build-docs` for a different change I ran into this causing
the command to fail. This change updates the docs so that they can build
properly again.