Extending the control plane model (#3938)
* The control plane model now offers a concise view of the Metastore.

The indexing scheduler query the metastore at all.

Other refactoring

We rely on SourceUid more broadly in the code.
The unit test were using IndexUid in a broken way in places. This is now
fixed.

This PR introduces a LogicalIndexTask. The idea is to have it hold all
information necessary for us to do the placement.
Right now this means only the max pipeline per index parameter (allowing
us to remove the ugly lookup in the source config map).

In the next PR we will most likely also break the 1:1 relationship
between LogicalIndexTask and IndexTask.

The idea would be to have a single logical index task, and possibly
break it into several IndexTask.

This PR also cleans up the hack the populates the list of indexing task
with ingest v2 tasks by mutating the list of tasks